update testscripts directory
1. add smoke.py
2. add socks.py
3. add tls.py
4. add browser.py
5. remove main.py
